fix(abrp): кастовать Float→Double для JSONObject.put на Android
NoSuchMethodError на устройстве: Android org.json.JSONObject не имеет
put(String, float) — только put(String, double). Десктопный org.json в
JVM unit-тестах имеет, поэтому регрессия пролезла.

Поля ChargingReading.chargingCapacityKwh и BatteryReading.sohPercent
типа Float — теперь явно .toDouble() перед put.

versionCode 290 → 291 (versionName остаётся 2.6.0) для пере-установки
как upgrade.

+++ app/src/main/kotlin/com/bydmate/app/data/remote/IternioTelemetryClient.kt
@@ -96,12 +96,16 @@ class IternioTelemetryClient @Inject constructor(
                     telemetry.put("is_dcfc", if (gun in DCFC_GUN_STATES) 1 else 0)
                 }
                 // -1.0f is the autoservice "no value" sentinel — drop it.
+                // .toDouble() is required: Android's JSONObject only exposes
+                // put(String, double) — there is no put(String, float). On JVM
+                // the desktop org.json has the float overload, so this lands as
+                // NoSuchMethodError only at runtime on the device.
                 c.chargingCapacityKwh?.takeIf { it >= 0f }?.let {
-                    telemetry.put("kwh_charged", it)
+                    telemetry.put("kwh_charged", it.toDouble())
                 }
             }
             battery?.sohPercent?.takeIf { it in 0f..100f }?.let {
-                telemetry.put("soh", it)
+                telemetry.put("soh", it.toDouble())
             }
 
             carModel?.trim()?.takeIf { it.isNotEmpty() }?.let { telemetry.put("car_model", it) }
